# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/notifications/{notificationId}/read": {
      "post": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/notifications/read-all": {
      "post": {
        "operationId": "NotificationsController_markAllRead",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/media/assets": {
      "post": {
        "operationId": "MediaAssetsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
    "/media/assets/{assetId}": {
      "get": {
        "operationId": "MediaAssetsController_get",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
    "/media/assets/{assetId}/clips": {
      "post": {
        "operationId": "MediaAssetsController_createClip",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
...
    "/fraud-cases/{caseId}/resolve": {
      "post": {
        "operationId": "FraudCasesController_resolve",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fraud"
        ]
      }
    },
    "/fraud-cases/{caseId}/dismiss": {
      "post": {
        "operationId": "FraudCasesController_dismiss",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fraud"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zeyoo API",
    "description": "Zeyoo backend public and internal API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterAssetDto": {
        "type": "object",
        "properties": {}
      },
      "CreateClipDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}